fix: start_rec_cb
start_rec callback function was not able to get ahold of the file created by output, possibly because the function fired the signal before the file was created.

I moved the job of finding the file (not the way I would want) to the file_changed callback function.

I also added a check for recording location inside of stop_rec callback function, to make sure that the script is at all times aware of the most recent file.