improv: added isRecording value to RecordingInfo
isRecording value is supposed to keep track of the current state of the recording.

If the user is recording it will not allow for change in title to avoid situation where the title changes and moves the recording to wrong place.

It will still allow a situation where user changes title by changing the hooked application