refactor: global variables, toned down calls for uuid
Moved some of the variables inside of classes to global scope instead, I will try to find some materials and ask around to see if I can improve it.

In previous version, the calls were move frequent, since it was located in both 'script_load()' and 'script_update()'.
After long thinking session I came to conclusion that the UUID fetch is unneccessary there, so as of now the UUID of selected source is refreshed in 5 cases:
1. User selected other source
2. User clicked button to see the UUID of selected source
3. User stopped recording
4. Recording was split either manually or by automatic recording splitter feature
5. User saved replay buffer

Note:
Callback function used by modified property event is shared between the source list and the UUID button.
I will have to think if I need to separate them.