feat: removed start signal in favour of event
Signals are not reliable enough for them to be used for initial script set-up.

Found by @Penwywern from OBS Discord server

Additionally I needed to create a fake buffer for the script, so the file movement wouldn't be blocked by "file is in use" situation, so I brought back old code from pre-optimization time as it works the best, although I really hate it.