feat: scene collection change triggers stop of recording/replay
OBS_FRONTEND_EVENT_SCENE_COLLECTION_CHANGING does not turn off the recording/replay buffer, but it does re-load the scripts, which in turn clears everything that starting recording/replay buffer collects when the start actually happens, so after stopping/saving the script is oblivious to what is the sourceUUID that it should use, what is the game title which will automatically send everything to Manual Recording in this case, fixed by adding an FINISHED_LOADING event will run the first hook_sh() and later the script will make sure that the recording/replay will stop automatically to force user to start only in one scene collection (to be changed to persisent value annotating program finishing loading).

TLDR; Changing Scene Collection un-loads the script, which deletes all data neccessary for organizing, so the script will check if recording/replay buffer is happening and turn it off if any of them are active (to be changed).